Continuing to work on this a bit. My brain has been chewing on this project in the back of my mind for the past week, so I am doing stuff based on a hodge podge of memories.
I believe there's still stuff I gotta do, namely testing for
exponentials. I know that that stuff is all in my crazy little head
and that there's no real right way to do it, so I think I'm going to
change PEMDAS to PEDMDAS by having the dice notation be the next in
line after parens and exponentials.

Aside from that, there are only two more things to do: clean up the
error messages so that they print pretty ones that explain exactly
what's wrong, like when there are missing parens or something is not
parsable. I should also finalize the documentation and get it up on
pypi as well.