Instead of rolling my own factorial function, I just used the built-in from the math lib as that appears to be following the same method that I was and disallowing negative factorials.
I am also currently attempting to graple with the order of operations
for dice notation and exponentials. At this point, I really just have
to choose one or the other. I believe that exponentials should be given
higher priority and that dice should come second based on how I was
intuiting it with my tests. This is something I am going to have to
document so that people are aware of this if they go to use it that,
if they want the dice to be recognized first, they must put it in
parens.