Re-doing underlying string parser.
Things are going well toward redoing the parser and moving
from parsley to pyparser. However, there are some edge
cases currently that are not going well, namely going back
and forth between the different order of operations.

There are also some errors that we're not throwing like
we used to. Things like "9." are getting parsed by the
new parser. Either we have to change our expectations or
change how the code interprets numbers, both int and floats.