Resolved issue with unary minus!
Resolved the issue where the unary minus was not working. Was due
to it not being implemented in the correct order of operations and
also some special logic needing to be added to the evaluator. The
problem is that the operator at the point that we're checking is no
longer a string but a function and I was checking against the string.