Adding more tests.
I am concerned about my order of operations, so I went
and got some equations from math websites with the answers
so that I could ensure that everything is working correctly.

I also noticed my coverage went down and it's because of click
so I added a base test for that before I fix the verbose output
issue so that I can get the numbers back up.