Woo! Version 2.0!
This is a big deal for me. I have just finished up the verbosity
issues that were the cause of the 2.0 rewrite and here we are!
The "-v" flag will now return dice rolls for all dice rolled. We
are now on a solid foundation to add any new functionality going
forward.

I also re-added mypy to my mix and tried to fix some of the
underlying type issues that were present.