feat(click): Added minimum dice roll option. test: Added tests for minimum dice roll flag.
While working on the minimum dice roll flag, I found that there
are a few tests that do not have properly set minimums. Worse yet,
the dice roller does not appear to handle <float>d<int> cases well
it seems and is giving values lower than the minimum that it should.