fix(diceparser): Fixed issue with float num_dice. test: Added further tests to ensure float num_dice values work.
We had an issue where, if num_dice was given as a float, we were
making changes to all dice rolls instead of the last one that we
needed to make which was causing our rolls to not have the correct
percentage chances. This has been corrected and the fractional handler
will only be applied at the end after all other rolls have been made.