feat(pyparser) Adding "k" keep notation.
Nearly done on keep notation, but there are a few edge cases that
I at least need to test for and figure out the behavior for. There
are things like "(4d6 + 10d4)k4" that would potentially work but
I am unsure what it will do and what the correct outcome should
be in those cases. I may have it such that the last operation must
have been a dice operation in order for the dice modifier operations
to work.