Continuing work on the keep notation. Before venturing down this rabbit hole, I was not aware just how much was going to need to change in order to have this option in place.
The main issue appears to be that the keep notation requires an implied
dice having been rolled beforehand as the last operation. In order to
capture this (as well as the rolls that make up the sum), it required
additional changes to the flow of the parser that we had in place. There
were a few options to attempt to deal with this:

1) Keep most things the same but make it so that, if we have a dice to
roll but it includes "k/K int?" at the end, then we would parse it a
different way.
  - This seemed easier but not the best way to handle this. Our parsing
  lines are already overly verbose and we're getting recursion check
  errors already by pyparsing so this scared me away from doing it.
2) Create a new object type that would represent the result of an
expression that was evaluated during the process of parsing an entire
expression. In this way the history of what was parsed and what order
could be preserved along with all dice rolls in such a way that we could
apply the keep notation to it. This was also going to be compatible with
all new dice modifications and would open the door so that we could add
them with less difficulty and no recursion warnings/errors.
  - While I went with this, it's requiring a considerable amount of
  changes in order to apply. I have written and rewritten changes in
  order to attempt to add this. I think the main issue I am having right
  now is that I am testing all of the tests we have but I'm attempting
  to work using the stable ground principle and it's causing me to have
  a hard time piecing together what works through all of the errors that
  we are getting.