Continuing to cut things up and fill out the operations. I'm nearly there with the operations, only got to do the finishing touches on the dice rolling functions so that they mesh with everything else that we're expecting.
I think that moving everything over to EvaluationResults objects is
really going to help with some of the things that are coming down the
pipeline, as well as making all of the typing easier to handle. All of
the veneer functions that are handling the operations are going to
greatly simplify the headache attempting to get mypy to agree that
everything is syncing correctly.

I still have not put in place the method in the parser that is going to
call the correct operations. What we're getting from the parser for any
order of operations level is a list of a list of ints, floats,
EvaluationResults, and strings representing the outcomes, inputs, and
operation types that we will be doing. e.g. [[1, "+", 1]]. This is easy
enough when considering a single operation, but multiple can be changed
like so: [[1, "+", 2, "+", 3]]. In order to handle this, I propose that
we solve the first pair first, then use that result with the second
given operation and so forth. I believe that we can use a for loop and
use the index skip to grab pairs (and be able to say that the value to
use with the next operation is op_index + 1)