Another big win. I was able to get the errors down substantially. Now there are only a handful left, mostly due to the additions made to the inFixNotation scheme that we have for the parser.
One issue that I do not know right away how to fix is going to be
re-adding the min and max roll options in a sane way to the parser. We
can do like we did before and have a global flag that can be used, but I
want to see if there is any way we can do it by passing additional info
into the pyparsing methods that we're calling.

We may have to move away from the inFixNotation that we have as it may
not encapsulate what we're doing well. The optional number for (1)d6 for
instance is causing a recursion issue how we currently have it stated.
It is likely causing the keep notation to not be able to parse strings
where the keep dice number is not supplied.