Addition and rolls verbose output working!
I have a working skeleton for what I need to do, now I need to make the
modifications to the other operations. Because this is going to require
a change in code across the different ops, I think the best thing to do
is create a wrapper function to gather the roll info as well as the
history at this point.