Removed old output test.
The new verbose output is going to include all of the info for what has
transpired to get us to the point where we have a value. This means that
checking whether the string'd output is the value will no longer work.

This test is removed and I will now need to make new tests that will
deal with the new verbose output. This will also allow for order of
operations testing which should be really helpful going forward when
adding more to our parser.

This really should have been done first. I don't know why I didn't write
tests first to ensure I knew what it was that I wanted to make before
changing any of the code. Lesson learned.