Getting close to finishing verbose printing!
The code appears to be done but further testing is required to ensure
that it is working properly. I should add in the tests using the
decorator so that I do not have to keep writing functions for it. We
need individual tests for the operations as well as combinations.

One thing that's worrying me however is that there appears to be a
problem with the order of operations for the roll "2+3-4*5/2". In this
case, it's doing the "2+3" addition first, then multiplying, dividing,
and finally subtracting. The expected order is for it to multiply,
divide, add, and then subtract. This isn't causing any order of ops issues
currently, and we have considerable testing to ensure that things are
being handled properly currently, but it does cause the history to be
out of order. I am not sure if this is simply a history issue or if it's
a problem with how the parser is handling things. I first thought that
maybe this had to do with the fact that the minus symbol is pulling
double duty as it could be being first parsed as a negative instead of
as an operation, but I have tried it with addition and substraction and
found the same issue. I do not know if this needs to be resolved right
away, but it should be investigated at some point. If I do not get it
done this time around I will need to make a ticket to track it.