Verbose printing done!
This concludes the verbose printing functionality additions. I have also
made some minor fixes. I am very happy with this addition as it was able
to increase readability and also help ensure our testing is top notch.

The issue with the order of operations with things like "3+2-1*7" still
persists. It's wanting to perform the addition first it seems and it
will take additional time to sort through. This is a parser issue
however and we're already considering changing to a new one anyway.