Mypy issues addressed!
I got all of the type issues fixed. I also was able to get the import
problem addressed as well. There's an explicit way that mypy wants
things exported within the __init__ files that I had to do.

At this point, the last thing that needs to be taken care of is the
precommit hook's issues. It found some whitespace it doesn't like and
black also came back on a few files. Will need to see how that works and
whether I can have it automatically change the code instead of simply
complaining about it.