Finally passed pre-commit!
I also figured out how to get it to automatically do the pre-commit as
well. This is awesome after I got over the initial hump of getting the
project up to the standards that they set out.

The next TODO items are going to be creating some documentation and also
redoing the README file to the new standard.