More fun changes.
The problem with pre-commit hooks is they only analyze the pieces of
code that are in the commit, not the entirety of the project. This makes
initial setup difficult because there may be files (or in my case, many
files) that are't touched that the pre-commit hooks would have issue
with if they were to run against them.