Removing __init__.py's from coverage report.
I keep getting the error message: `No source for code:
'/Users/runner/work/roll-cli/roll-cli/src/roll_cli/__init__.py'.`

No idea why, no real direction given as to why or which report doesn't
have this because this is a collection of all of the reports that have
been generated put together so I am guestimating that one of the
versions that is doing testing is not including it in its report. I have
tried to research it but we're not given the source files that it's
considering for the report, only the tests run from what I can tell. It
might be in the artifact file that's being transmitted after the run.