Removed some previous files, updated coverage options.
It appears that coverage.py does not like when you have multiple areas
that are specifying what to do, even if they are not contradictory
options. I had some things specified in a .coveragerc file, some in my
nox file, and some more in my pyproject.toml. It didn't like it and it
actually was causing my project to not have the correct coverage totals.

This appears to fix it, at least locally. I do not know whether my
pyproject.toml file is used in github actions, so this will be a test
commit to see.