Adding drop highest/lowest notation!
This was very straight forward to add. A lot of reuse out of the keep
notation components. The only thing that gave me trouble were the tests
were not a straight copy/paste from the keep notation since they are
essentially exact opposites of each other.

The other thing is that we may want to move both the keep and the drop
notation items to the operations file. I am not sure why I kept the keep
notation in the diceparser, but I think it's because it's dealing with
toks and not straight dice object things.