Updating tests for notation edge cases.
We currently have an issue with the order of operations when it comes to
the dice notation against itself. I believe that it should probably
ready left-to-right, but, instead, it's doing things piecemeal. I am not
entirely certain the best course of action because this is the part I'm
essentially making up as I go, but it feels right to me that these
things, including the extra notation such as keep and drop, should all
be read from left-to-right. I will create a bug fix ticket for it and
work on that logic change separately however.

I also noticed that `4d6X1d2` is stating that it is performing an add?