Added comparison operators!
Ran into a problem with the Maximum roll options however for the roll
function. I think it just might not work at that point and if I were to
instantiate a new DiceParser object then maybe it would? Either way, I
should make a ticket to look into it further. I think the solution that
I came up with is pretty dang hacky. I understand that I was trying to
get around the limitation that we're getting string input but I believe
the real solution is that those commandline options needs to be parsed
at the __main__.py level and then all options should be passed down as
parameters from there on out.

I believe the problem was that I couldn't pass additional info to the
Parser object and that's why I was keeping state with a global variable?
Their 3.0.0 version just came out anyway, might as well see if they have
something for this as well.