Finished making nox happy!
Had a few setbacks, but was able to figure it out in the end. Right now,
because we're using `__eq__` for a number of things, I cannot use it as
I am using the other comparison method overloads. In the future, I may
want to consider changing those other methods to only do the comparison,
especially as I am not currently using them correctly per the official
guidelines because I am not taking any object nor returning a boolean.
For some reason I wasn't get the same alerts as I did for `__eq__`, but
I am going to assume that doesn't mean I shouldn't also follow the same
rules in those cases.

One thing that I have been mulling over is a refactor for the DiceParser
anyway that would allow me to remove a considerable amount of reused
boiler plate code. We have the initial get-left-hand-value and then the
for-loop that then allows us to use the next operation and
left-hand-value, but that means then that we should make functions to
handle those operations and use it inside of a helper method that does
that heavy lifting only in one place.